Added available models as choices for the model cli arg to make it so that users shouldn't provide a model that doesn't exist
